// client/library-app/src/models/LoanRecord.ts
import { Book } from "./Book";

export type LoanRecord = {
    _id: string;
    status: "LOANED" | "AVAILABLE";
    loanedDate: Date;
    dueDate: Date;
    returnedDate?: Date;
    student: string;
    employeeOut: string;
    employeeIn?: string;
    item: Book;
    createdAt: Date;
    updatedAt: Date;
}